Move mech submodule, add test (#94)
* Move mech submodule, add test
