+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "prediction_market_agent/agents/microchain_agent/mech/mech"]
-	path = prediction_market_agent/agents/microchain_agent/mech/mech
+[submodule "prediction_market_agent/tools/mech/mech"]
+	path = prediction_market_agent/tools/mech/mech
 	url = https://github.com/valory-xyz/mech.git
