+++ mypy.ini
@@ -15,7 +15,7 @@ disallow_incomplete_defs = True
 disallow_untyped_defs = True
 ignore_missing_imports = True
 # Exclude submodules that are themselves not type-checked
-exclude = prediction_market_agent/agents/microchain_agent/mech/mech/
+exclude = prediction_market_agent/tools/mech/mech/
 
 # See https://github.com/python/mypy/issues/3905#issuecomment-421065323
 # We don't want to ignore all missing imports as we want to catch those in our own code
