+++ prediction_market_agent/tools/mech/api_keys.py
@@ -1,0 +1,25 @@
+import typing as t
+
+from prediction_market_agent_tooling.tools.utils import check_not_none
+from pydantic import SecretStr
+
+from prediction_market_agent.utils import APIKeys
+
+
+class MechAPIKeys(APIKeys):
+    GOOGLE_SEARCH_API_KEY: t.Optional[SecretStr] = None
+    GOOGLE_SEARCH_ENGINE_ID: t.Optional[SecretStr] = None
+
+    @property
+    def google_search_api_key(self) -> SecretStr:
+        return check_not_none(
+            self.GOOGLE_SEARCH_API_KEY,
+            "GOOGLE_SEARCH_API_KEY missing in the environment.",
+        )
+
+    @property
+    def google_search_engine_id(self) -> SecretStr:
+        return check_not_none(
+            self.GOOGLE_SEARCH_ENGINE_ID,
+            "GOOGLE_SEARCH_ENGINE_ID missing in the environment.",
+        )
